Separate out another advance-past-constituent function which will be reused in other oracles Also separate out a find-previous function
