+++ stanza/models/constituency/dynamic_oracle.py
@@ -16,6 +16,48 @@ def advance_past_constituents(gold_sequence, cur_index):
         cur_index = cur_index + 1
     return None
 
+def find_previous_open(gold_sequence, cur_index):
+    """
+    Go backwards from cur_index to find the open which opens the previous block of stuff.
+
+    Return None if it can't be found.
+    """
+    count = 0
+    cur_index = cur_index - 1
+    while cur_index >= 0:
+        if isinstance(gold_sequence[cur_index], OpenConstituent):
+            count = count + 1
+            if count > 0:
+                return cur_index
+        elif isinstance(gold_sequence[cur_index], CloseConstituent):
+            count = count - 1
+        cur_index = cur_index - 1
+    return None
+
+def find_in_order_constituent_end(gold_sequence, cur_index):
+    """
+    Advance cur_index through gold_sequence until the next block has ended
+
+    This is different from advance_past_constituents in that it will
+    also return when there is a Shift when count == 0.  That way, we
+    return the first block of things we know attach to the left
+    """
+    count = 0
+    saw_shift = False
+    while cur_index < len(gold_sequence):
+        if isinstance(gold_sequence[cur_index], OpenConstituent):
+            count = count + 1
+        elif isinstance(gold_sequence[cur_index], CloseConstituent):
+            count = count - 1
+            if count == -1: return cur_index
+        elif isinstance(gold_sequence[cur_index], Shift):
+            if saw_shift and count == 0:
+                return cur_index
+            else:
+                saw_shift = True
+        cur_index = cur_index + 1
+    return None
+
 class DynamicOracle():
     def __init__(self, root_labels, oracle_level, repair_types, additional_levels, deactivated_levels):
         self.root_labels = root_labels
